CRT-846 Band highlight with keyboard navigation
